Default

+++ detect.py
@@ -33,9 +33,11 @@ def main():
         t1 = time.time()
         img = img[np.newaxis, :, :, :]
         loss, detections = ano.detect(img, './params.yaml')
-        detections = ((detections[0, :, :, 0] + 1.) * 255. / 2.).astype(np.uint8)
-        img = (img[0, :, :, 0] * 255.).astype(np.uint8)
+        detections = (detections[0, :, :, 0] + 1.) * 255. / 2.
+        img = img[0, :, :, 0] * 255.
         diff = np.abs(img - detections).astype(np.uint8)
+        img = img.astype(np.uint8)
+        detections = detections.astype(np.uint8)
         img_name = str_num(i + 1, 2) + '_' + str(lbl) + '.png'
         dt_name = str_num(i + 1, 2) + '_' + str(lbl) + '_dtc.png'
         diff_name  = str_num(i + 1, 2) + '_' + str(lbl) + '_diff.png'
